src/main/java/dev/ardijorganxhi/springboottemplate/controller/AuthController.java
package dev.ardijorganxhi.springboottemplate.controller;


import dev.ardijorganxhi.springboottemplate.model.request.LoginRequest;
import dev.ardijorganxhi.springboottemplate.model.request.RegisterRequest;
import dev.ardijorganxhi.springboottemplate.service.AuthService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/auth")
@RequiredArgsConstructor
public class AuthController {

    private final AuthService authService;

    @PostMapping("/register")
    public void register(@RequestBody RegisterRequest request) {
        authService.register(request);
    }

    @PostMapping("/login")
    public ResponseEntity<String> login(@RequestBody LoginRequest request) {
        return ResponseEntity.ok(authService.login(request));
    }
}
